refactor(core): move KvBackend trait + version-stamp retry into solobase-core
solobase-cloudflare is wasm32-only and excluded from cargo test
--workspace, so put_version_stamp_with_retry had no host-runnable
tests (and kv_cached_db's trait doc referenced a tests/support
MockKvBackend that never existed). Follow the cache_key extraction
precedent: the KvBackend trait (wasm32/native async_trait cfg_attrs
kept) and the retry helper move to solobase_core::kv; WorkerKvBackend
and KvCachedD1DatabaseService stay in solobase-cloudflare, whose
callers now import from core (no compat shim).
